output overall status
