+++ client/src/components/molecules/VersionNumbers/VersionTable.module.scss
@@ -56,21 +56,3 @@
         }
     }
 }
-
-.spinner {
-    composes: spinner, small from '../../atoms/Spinner.module.scss';
-    margin-right: $spacer;
-}
-
-.handle {
-    display: inline-block;
-    border: 0;
-    background: none;
-    box-shadow: none;
-    padding: 0;
-    margin: 0;
-    margin-left: -1rem;
-    padding-right: .5rem;
-    cursor: pointer;
-    color: $brand-grey-light;
-}
