--- client/src/components/molecules/VersionNumbers/utils.test.ts
@@ -1,11 +1,0 @@
-import { isJsonString } from './utils'
-
-describe('isJsonString', () => {
-    it('detects json correctly', () => {
-        const testJson = isJsonString('{ "hello": "squid" }')
-        expect(testJson).toBeTruthy()
-
-        const testNonJson = isJsonString('<strong>')
-        expect(testNonJson).toBeFalsy()
-    })
-})
