--- client/src/components/molecules/VersionNumbers/utils.ts
@@ -1,8 +1,0 @@
-export function isJsonString(str: string) {
-    try {
-        JSON.parse(str)
-    } catch (e) {
-        return false
-    }
-    return true
-}
